commit source

+++ 05_Source/OZTT_Phone/src/main/java/com/org/oztt/controller/ShopCartController.java
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.alibaba.fastjson.JSONObject;
 import com.org.oztt.contants.CommonConstants;
 import com.org.oztt.formDto.ContCartItemDto;
-import com.org.oztt.formDto.ContCartItemListDto;
 import com.org.oztt.formDto.ContCartProItemDto;
 import com.org.oztt.service.GoodsService;
 
@@ -38,11 +37,15 @@ public class ShopCartController extends BaseController {
     public String init(Model model, HttpServletResponse response, HttpSession session) {
         try {
             String customerNo = (String) session.getAttribute(CommonConstants.SESSION_CUSTOMERNO);
+            if (StringUtils.isEmpty(customerNo)) {
+                return "redirect:/login/init";
+            }
+
             // 登陆成功以后取得购物车中的数据然后更新Cookie
             goodsService.deleteCanNotBuyGoodsByCustomer(customerNo);
             List<ContCartItemDto> consCarts = goodsService.getAllContCartForCookie(customerNo);
 
-            List<ContCartItemListDto> cartsList = new ArrayList<ContCartItemListDto>();
+            //List<ContCartItemListDto> cartsList = new ArrayList<ContCartItemListDto>();
             if (!CollectionUtils.isEmpty(consCarts)) {
                 for (ContCartItemDto dto : consCarts) {
                     if (StringUtils.isEmpty(dto.getGoodsPropertiesDB())) {
@@ -58,57 +61,19 @@ public class ShopCartController extends BaseController {
                             new BigDecimal(dto.getGoodsQuantity()), 2, BigDecimal.ROUND_DOWN)));
                 }
 
-                // 取得购物车天数数据
-                //String[] prostr = getShopCartPro();
+                //ContCartItemListDto contCartItemListDto = new ContCartItemListDto();
 
-                ContCartItemListDto contCartItemListDto = new ContCartItemListDto();
-                //List<ContCartItemDto> timeDto = new ArrayList<ContCartItemDto>();
-                //int q = 0;
-                //                for (int i = 0; i < prostr.length; i++) {
-                //                    contCartItemListDto = new ContCartItemListDto();
-                //                    timeDto = new ArrayList<ContCartItemDto>();
-                //                    for (ContCartItemDto dto : consCarts) {
-                //                        if (i == 0) {
-                //                            if (Integer.valueOf(prostr[i]) >= Integer.valueOf(dto.getCanbuyDay())) {
-                //                                timeDto.add(dto);
-                //                                q++;
-                //                            }
-                //                        }
-                //                        else {
-                //                            if (Integer.valueOf(prostr[i - 1]) < Integer.valueOf(dto.getCanbuyDay())
-                //                                    && Integer.valueOf(prostr[i]) >= Integer.valueOf(dto.getCanbuyDay())) {
-                //                                timeDto.add(dto);
-                //                                q++;
-                //                            }
-                //                        }
-                //
-                //                    }
-                //                    if (timeDto.size() > 0) {
-                //                        contCartItemListDto.setQueryDay(prostr[i]);
-                //                        contCartItemListDto.setItemList(timeDto);
-                //                        cartsList.add(contCartItemListDto);
-                //                    }
-                //
-                //                }
-                // 将剩余的购物的放入list中
-                //                contCartItemListDto = new ContCartItemListDto();
-                //                timeDto = new ArrayList<ContCartItemDto>();
-                //                for (int i = 0; i < consCarts.size(); i++) {
-                //                    if (i >= q) {
-                //                        timeDto.add(consCarts.get(i));
-                //                    }
-                //                }
 
-                if (consCarts.size() > 0) {
-                    //contCartItemListDto.setQueryDay(super.getApplicationMessage("shopcart_other_time", session));
-                    contCartItemListDto.setItemList(consCarts);
-                    cartsList.add(contCartItemListDto);
-                }
-
-                model.addAttribute("count", consCarts.size());
+//                if (consCarts.size() > 0) {
+//
+//                    contCartItemListDto.setItemList(consCarts);
+//                    cartsList.add(contCartItemListDto);
+//                }
+//
+//                model.addAttribute("count", consCarts.size());
             }
 
-            model.addAttribute("cartsList", cartsList);
+            model.addAttribute("cartsList", consCarts);
             return "shopcart";
         }
         catch (Exception e) {
